Try do calculation, need to fix

+++ src/main/java/com/feie/calculator/models/TaxResults.java
@@ -1,34 +1,32 @@
 package com.feie.calculator.models;
 
-import java.math.BigDecimal;
-
 public class TaxResults {
-    private BigDecimal incomeAfterFEIE;
-    private BigDecimal federalTax;
-    private BigDecimal incomeAfterTax;
+    private double incomeAfterFEIE;
+    private double federalTax;
+    private double incomeAfterTax;
 
     // Getters and setters
-    public BigDecimal getIncomeAfterFEIE() {
+    public double getIncomeAfterFEIE() {
         return incomeAfterFEIE;
     }
 
-    public void setIncomeAfterFEIE(BigDecimal incomeAfterFEIE) {
+    public void setIncomeAfterFEIE(double incomeAfterFEIE) {
         this.incomeAfterFEIE = incomeAfterFEIE;
     }
 
-    public BigDecimal getFederalTax() {
+    public double getFederalTax() {
         return federalTax;
     }
 
-    public void setFederalTax(BigDecimal federalTax) {
+    public void setFederalTax(double federalTax) {
         this.federalTax = federalTax;
     }
 
-    public BigDecimal getIncomeAfterTax() {
+    public double getIncomeAfterTax() {
         return incomeAfterTax;
     }
 
-    public void setIncomeAfterTax(BigDecimal incomeAfterTax) {
+    public void setIncomeAfterTax(double incomeAfterTax) {
         this.incomeAfterTax = incomeAfterTax;
     }
 }
